Remove files if exit status -eq 0

+++ functions.bash
@@ -8,7 +8,7 @@
 wav_flac ()
 {
     flac --keep-foreign-metadata --best --force --silent "$1".wav
-    if [ flac --test "$1".flac ]
+    if [ `flac --test "$1".flac` ]
     then
         rm "$1".wav
     fi
@@ -22,16 +22,30 @@ ogg_encode ()
 mp3_ogg ()
 {
     lame --decode "$1".mp3
-    rm "$1".mp3
+    if [ $? -eq 0 ]
+    then
+        rm "$1".mp3
+    fi
+    
     ogg_encode "$1".wav
-    rm "$1".wav
+    if [ $? -eq 0 ]
+    then
+        rm "$1".wav
+    fi
 }
 
 wma_ogg ()
 {
     mplayer -quiet -ao pcm:file="$1".wav -vc null -vo null "$1".wma
-    rm "$1".wma
+    if [ $? -eq 0 ]
+    then
+        rm "$1".wma
+    fi
+    
     ogg_encode "$1".wav
-    rm "$1".wav
+    if [ $? -eq 0 ]
+    then
+        rm "$1".wav
+    fi
 }
